abstract test comp
